# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Open Agents HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
  /api/v1/agents:
    get:
      operationId: listAgents
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAgentsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Return cached supported and locally installed agent adapters
      tags:
      - agents
  /api/v1/agents/{agent}/auth:
    post:
      operationId: startAgentAuth
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartAgentAuthResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Open the fixed native authentication flow for one agent
      tags:
      - agents
...
  /api/v1/agents/{agent}/models:
    get:
      operationId: getAgentModels
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      - description: Optional project identifier used as the model-catalog cache scope.
        in: query
        name: projectId
        schema:
          description: Optional project identifier used as the model-catalog cache
            scope.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentModelsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Return the cached model picker for one agent, discovering it on first
        use
      tags:
      - agents
  /api/v1/agents/{agent}/models/refresh:
    post:
      operationId: refreshAgentModels
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      - description: Optional project identifier used as the model-catalog cache scope.
        in: query
        name: projectId
        schema:
          description: Optional project identifier used as the model-catalog cache
            scope.
          type: string
      - description: When true, compare executable and config metadata before running
          discovery.
        in: query
        name: revalidate
        schema:
          description: When true, compare executable and config metadata before running
            discovery.
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentModelsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Refresh and cache the model picker for one agent
      tags:
      - agents
  /api/v1/agents/{agent}/probe:
    post:
      operationId: probeAgent
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbeAgentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Ensure launch-fresh readiness for one agent adapter
      tags:
      - agents
...
  /api/v1/agents/auth-plans:
    get:
      operationId: listAgentAuthPlans
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAgentAuthPlansResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Return display-safe native authentication plans for supported agents
      tags:
      - agents
...
  /api/v1/agents/readiness:
    get:
      operationId: getAgentReadiness
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentReadinessResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Return cached normalized agent readiness without running native checks
      tags:
      - agents
  /api/v1/agents/readiness/ensure:
    post:
      operationId: ensureAgentReadiness
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnsureAgentReadinessRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentReadinessResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Ensure normalized readiness for selected agent adapters
      tags:
      - agents
  /api/v1/agents/refresh:
    post:
      operationId: refreshAgents
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAgentsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Refresh the cached local agent adapter catalog
      tags:
      - agents
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AgentAuthPlan:
      properties:
        action:
          type: string
        agentId:
          type: string
        available:
          type: boolean
        displayCommand:
          type: string
        documentationUrl:
          type: string
        guidance:
          type: string
        launchMode:
          enum:
          - terminal
          - documentation
          type: string
        reason:
          type: string
      required:
      - agentId
      - action
      - launchMode
      - available
      - documentationUrl
      type: object
    AgentAuthenticationObservation:
      properties:
        attemptedAt:
          format: date-time
          type:
          - "null"
          - string
        checkedAt:
          format: date-time
          type:
          - "null"
          - string
        freshness:
          enum:
          - fresh
          - stale
          - checking
          type: string
        reason:
          type: string
        reasonCode:
          type: string
        state:
          enum:
          - authorized
          - unauthorized
          - unknown
          - not_applicable
          type: string
      required:
      - state
      - freshness
      - checkedAt
      - attemptedAt
      - reasonCode
      - reason
      type: object
...
    AgentInfo:
      properties:
        authStatus:
          description: Advisory local auth probe result. authorized means a recent
            local probe passed; spawn remains the authoritative validation point.
          enum:
          - authorized
          - unauthorized
          - unknown
          type: string
        id:
          type: string
        label:
          type: string
        lastUsedAt:
          description: Creation time of the newest retained session currently attributed
            to this agent.
          format: date-time
          type:
          - "null"
          - string
        usageCount:
          description: Number of retained sessions currently attributed to this agent.
          type: integer
      required:
      - id
      - label
      type: object
...
    AgentInstallationObservation:
      properties:
        attemptedAt:
          format: date-time
          type:
          - "null"
          - string
        checkedAt:
          format: date-time
          type:
          - "null"
          - string
        freshness:
          enum:
          - fresh
          - stale
          - checking
          type: string
        reason:
          type: string
        reasonCode:
          type: string
        state:
          enum:
          - installed
          - not_installed
          - unknown
          type: string
      required:
      - state
      - freshness
      - checkedAt
      - attemptedAt
      - reasonCode
      - reason
      type: object
...
    AgentModelInfo:
      properties:
        defaultEffort:
          type: string
        efforts:
          items:
            type: string
          type: array
        id:
          type: string
        isDefault:
          type: boolean
        label:
          type: string
        provider:
          type: string
      required:
      - id
      - label
      type: object
    AgentModelsResponse:
      properties:
        agentId:
          type: string
        allowCustom:
          type: boolean
        binaryVersion:
          type: string
        customModelEntry:
          enum:
          - none
          - direct
          - configured
          type: string
        fetchedAt:
          format: date-time
          type: string
        models:
          items:
            $ref: '#/components/schemas/AgentModelInfo'
          type: array
        refreshRecommended:
          type: boolean
        selectionMode:
          enum:
          - catalog
          - text
          - mode
          type: string
        source:
          type: string
        stale:
          type: boolean
        validatedAt:
          format: date-time
          type: string
        warning:
          type: string
      required:
      - agentId
      - selectionMode
      - models
      - customModelEntry
      - allowCustom
      - source
      - fetchedAt
      - stale
      type: object
    AgentReadinessResponse:
      properties:
        agents:
          items:
            $ref: '#/components/schemas/AgentReadinessSnapshot'
          type: array
      required:
      - agents
      type: object
    AgentReadinessSnapshot:
      properties:
        authentication:
          $ref: '#/components/schemas/AgentAuthenticationObservation'
        effectiveReadiness:
          enum:
          - ready
          - not_ready
          - unknown
          type: string
        id:
          type: string
        installation:
          $ref: '#/components/schemas/AgentInstallationObservation'
        label:
          type: string
        lastUsedAt:
          format: date-time
          type:
          - "null"
          - string
        usageCount:
          type: integer
      required:
      - id
      - label
      - installation
      - authentication
      - effectiveReadiness
      - usageCount
      type: object
...
    EnsureAgentReadinessRequest:
      properties:
        agentIds:
          items:
            type: string
          type: array
        purpose:
          enum:
          - display
          - launch
          type: string
      required:
      - purpose
      type: object
...
    ListAgentAuthPlansResponse:
      properties:
        plans:
          items:
            $ref: '#/components/schemas/AgentAuthPlan'
          type: array
      required:
      - plans
      type: object
    ListAgentsResponse:
      properties:
        authorized:
          description: Compatibility list of installed agents whose local auth probe
            recently returned authorized. Advisory and stale-prone; spawn may still
            fail.
          items:
            $ref: '#/components/schemas/AgentInfo'
          type: array
        installed:
          description: Agents whose binary resolved during the latest best-effort
            local catalog probe.
          items:
            $ref: '#/components/schemas/AgentInfo'
          type: array
        supported:
          description: Agents supported by this daemon build.
          items:
            $ref: '#/components/schemas/AgentInfo'
          type: array
      required:
      - supported
      - installed
      - authorized
      type: object
...
    ProbeAgentResponse:
      properties:
        agent:
          $ref: '#/components/schemas/AgentInfo'
        installed:
          type: boolean
        supported:
          type: boolean
      required:
      - agent
      - supported
      - installed
      type: object
...
    ShellTerminalResponse:
      properties:
        createdAt:
          format: date-time
          type: string
        handleId:
          type: string
        projectId:
          type: string
        sessionId:
          type: string
        title:
          type: string
        workingDir:
          type: string
      required:
      - handleId
      - workingDir
      - title
      - createdAt
      type: object
...
    StartAgentAuthResponse:
      properties:
        action:
          type: string
        agentId:
          type: string
        guidance:
          type: string
        terminal:
          $ref: '#/components/schemas/ShellTerminalResponse'
        terminalInput:
          type: string
      required:
      - agentId
      - action
      - terminal
      type: object
...